Extract common logic into a new component and fix styling problems
