+++ src/components/project-page/ArtifactList.js
@@ -2,16 +2,8 @@
 
 import React from 'react';
 
-import {
-    Grid,
-    Typography,
-    List,
-    Paper,
-    Divider,
-    withStyles,
-} from '@material-ui/core';
-
 import ArtifactListItem from './ArtifactListItem';
+import EntityList from './EntityList';
 
 import { Artifact } from '../../api/models';
 
@@ -19,34 +11,15 @@ type Props = {
     artifacts: Array<Artifact>,
 };
 
-const styles = theme => ({
-    title: {
-        marginTop: theme.spacing.unit,
-        padding: theme.spacing.unit,
-    },
-});
-
 const ArtifactList = ({ artifacts }: Props) => {
     const artifactListItems = artifacts
         .map(artifact => <ArtifactListItem artifact={artifact} />);
 
     return (
-        <Grid item xs={6}>
-            <Paper elevation={2}>
-                <Typography
-                  variant="title"
-                  align="center"
-                >
-                    ARTIFACT
-                </Typography>
-                <Divider />
-                <List>
-                    {artifactListItems}
-                </List>
-            </Paper>
-        </Grid>
+        <EntityList title="ARTIFACTS">
+            {artifactListItems}
+        </EntityList>
     );
 };
 
-
-export default withStyles(styles)(ArtifactList);
+export default ArtifactList;
